+++ benchmark/inference/inference_benchmark.py
@@ -15,9 +15,18 @@ supported_sets = {
 }
 
 
+@torch.no_grad()
+def full_batch_inference(model, data):
+    model.eval()
+    y = model(data.x, data.edge_index)
+    return y
+
+
 def run(args: argparse.ArgumentParser) -> None:
 
-    device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
+    # cuda device is not suitable for full batch mode
+    device = torch.device(
+        'cuda' if not args.full_batch and torch.cuda.is_available() else 'cpu')
 
     print('BENCHMARK STARTS')
     for dataset_name in args.datasets:
@@ -56,14 +65,16 @@ def run(args: argparse.ArgumentParser) -> None:
                 print(f'Configuration of {dataset_name} + {model_name} '
                       f'not supported. Skipping.')
                 continue
+            with_loader = not args.full_batch or (model_name == 'pna'
+                                                  and degree is None)
             print(f'Evaluation bench for {model_name}:')
 
             for batch_size in args.eval_batch_sizes:
                 num_nodes = data[
                     'paper'].num_nodes if hetero else data.num_nodes
                 sampler = torch.utils.data.RandomSampler(
-                    range(num_nodes), num_samples=args.num_steps *
-                    batch_size) if args.num_steps != -1 else None
+                    range(num_nodes), num_samples=args.num_steps * batch_size
+                ) if args.num_steps != -1 and with_loader else None
 
                 if not hetero:
                     subgraph_loader = NeighborLoader(
@@ -74,8 +85,8 @@ def run(args: argparse.ArgumentParser) -> None:
                         shuffle=False,
                         num_workers=args.num_workers,
                         sampler=sampler,
-                    )
-                    if args.evaluate:
+                    ) if with_loader else None
+                    if args.evaluate and not args.full_batch:
                         test_loader = NeighborLoader(
                             data,
                             num_neighbors=[-1],  # layer-wise inference
@@ -98,7 +109,7 @@ def run(args: argparse.ArgumentParser) -> None:
                             shuffle=False,
                             num_workers=args.num_workers,
                             sampler=sampler,
-                        )
+                        ) if with_loader else None
 
                     for hidden_channels in args.num_hidden_channels:
                         print('----------------------------------------------')
@@ -132,9 +143,11 @@ def run(args: argparse.ArgumentParser) -> None:
                         model.eval()
 
                         # Define context manager parameters:
-                        cpu_affinity = subgraph_loader.enable_cpu_affinity(
-                            args.loader_cores
-                        ) if args.cpu_affinity else nullcontext()
+                        if args.cpu_affinity and with_loader:
+                            cpu_affinity = subgraph_loader.enable_cpu_affinity(
+                                args.loader_cores)
+                        else:
+                            cpu_affinity = nullcontext()
                         profile = torch_profile(
                         ) if args.profile else nullcontext()
                         itt = emit_itt(
@@ -142,16 +155,28 @@ def run(args: argparse.ArgumentParser) -> None:
 
                         with cpu_affinity, amp, timeit() as time:
                             for _ in range(args.warmup):
-                                model.inference(subgraph_loader, device,
-                                                progress_bar=True)
-                            time.reset()
+                                if args.full_batch:
+                                    full_batch_inference(model, data)
+                                else:
+                                    model.inference(subgraph_loader, device,
+                                                    progress_bar=True)
+                            if args.warmup > 0:
+                                time.reset()
                             with itt, profile:
-                                y = model.inference(subgraph_loader, device,
-                                                progress_bar=True)
-                            if args.evaluate:
-                                test_acc = model.test(y, test_loader, device,
-                                                      progress_bar=True)
-                                print(f'Test Accuracy: {test_acc:.4f}')
+                                if args.full_batch:
+                                    y = full_batch_inference(model, data)
+                                    if args.evaluate:
+                                        mask = data.test_mask
+                                        pred = y[mask].argmax(1)
+                                        test_acc = pred.eq(data.y[mask]).sum().item() / mask.sum().item()
+                                        print(f'Full Batch Test Accuracy: {test_acc:.4f}')
+                                else:
+                                    y = model.inference(subgraph_loader, device,
+                                                    progress_bar=True)
+                                    if args.evaluate:
+                                        test_acc = model.test(y, test_loader, device,
+                                                            progress_bar=True)
+                                        print(f'Mini Batch Test Accuracy: {test_acc:.4f}')
 
                         if args.profile:
                             rename_profile_file(model_name, dataset_name,
@@ -197,10 +222,11 @@ if __name__ == '__main__':
     add('--vtune-profile', action='store_true')
     add('--bf16', action='store_true')
     add('--cpu-affinity', action='store_true',
-        help="Use DataLoader affinitzation.")
+        help='Use DataLoader affinitzation.')
     add('--loader-cores', nargs='+', default=[], type=int,
         help="List of CPU core IDs to use for DataLoader workers.")
     add('--measure-load-time', action='store_true')
+    add('--full-batch', action='store_true', help='Use full batch mode.')
     add('--evaluate', action='store_true')
     add('--ckpt_path', type=str,
         help='checkpoint path to look for the model dict')